+++ tests/testthat/test-60-metamake-bexltall.R
@@ -1,4 +1,4 @@
 # BEXLTALL runs in its own file because it takes a long time
-test_that("px file = pxsave(px(px file))", {
-  expect_px_pxsave_preserves_everything("BEXLTALL")
+test_that("px file = px_save(px(px file))", {
+  expect_px_px_save_preserves_everything("BEXLTALL")
 })