+++ tests/testthat/test-71-px-pxave-equivalent.R
@@ -1,16 +1,16 @@
-test_that("px file and pxsave(px(px file), path='*.px') are equivalent", {
+test_that("px file and px_save(px(px file), path='*.px') are equivalent", {
   # Some files don't give an exact match in the px-file because there can be
   # subtle differences in formatting. PxJob is run on these file to test
   # that they are equivalent.
   skip_if_not_installed("pxjob64Win", minimum_version = "1.1.0")
 
-  run_px_pxsave_and_expect_equal <- function(table_name) {
+  run_px_px_save_and_expect_equal <- function(table_name) {
     px_in  <- get_px_file_path(table_name)
     px_out <- temp_px_file()
 
     px_in %>%
       px() %>%
-      pxsave(path = px_out)
+      px_save(path = px_out)
 
     pxjob_clean(input = px_in,  output = pxjob_in  <- temp_px_file())
     pxjob_clean(input = px_out, output = pxjob_out <- temp_px_file())
@@ -41,10 +41,10 @@ test_that("px file and pxsave(px(px file), path='*.px') are equivalent", {
     expect_equal(px_file_as_lines(pxjob_in), px_file_as_lines(pxjob_out))
   }
 
-  run_px_pxsave_and_expect_equal("BEXSTA_windows_1252")
-  run_px_pxsave_and_expect_equal("CONTVARIABLE")
-  run_px_pxsave_and_expect_equal("CONTVARIABLE_multiple_languages")
-  run_px_pxsave_and_expect_equal("no_timeval_or_codes2")
-  run_px_pxsave_and_expect_equal("SOXATI4")
-  run_px_pxsave_and_expect_equal("TUX01")
+  run_px_px_save_and_expect_equal("BEXSTA_windows_1252")
+  run_px_px_save_and_expect_equal("CONTVARIABLE")
+  run_px_px_save_and_expect_equal("CONTVARIABLE_multiple_languages")
+  run_px_px_save_and_expect_equal("no_timeval_or_codes2")
+  run_px_px_save_and_expect_equal("SOXATI4")
+  run_px_px_save_and_expect_equal("TUX01")
 })